Replace the old match checking algorithm
